[OPS-1002] [OPS-1004] Add handling for Scala literals and JAAS settings. Experimental support for APPEND settings via env vars.
This commit implements the first part of OPS-1002 by introducing special
handling for environment variables that are literals. It detects them either via
OPTS_LITERAL, or if they include the keywork SASL_JAAS_SETTINGS.  For such
variables we try to strip leading and trailing double quotes, then leading and
trailing single quotes, then enclose them in three double quotes.

This commit implements introduces the EXPERIMENTAL flag. If it is set,
experimental features are used.

This commit implements OPS-1004 as an experimental feature. If it detects
LENSES_APPEND_CONF or SECURITY_APPEND_CONF, it will append their configs
to the respective files. The EXPERIMENTAL flag must be set for this
functionality.

Signed-off-by: Marios Andreopoulos <opensource@andmarios.com>