Bugfix: delete /data/security.conf on start
This is the same behaviour we follow for `/data/lenses.conf` but missed to do so
for `/data/security.conf`. These files are created automatically when the
container starts. We offer facilities (`/mnt/settings/lenses.conf`
and `/mnt/secrets/security.conf`) for people who want to set these files
themselves.

[OPS-2169]

Signed-off-by: Marios Andreopoulos <opensource@andmarios.com>