Switch to slim base image, remove checkport and wget
Checkport is not required, Lenses will fail itself if the needed ports are in
use.
Wget is not required, can be installed via DEBUG_TOOLS=1 if needed.

Furthermore checkport was the last arch-specific program. We do have arm
versions, but removing it completely unlocks ARM64 images without any special
code in the Dockerfile.

[OPS-2208][OPS-2159]

Signed-off-by: Marios Andreopoulos <opensource@andmarios.com>